requirements.txt
streamlit
pyrebase4
google-generativeai
